imap: refactor Account/Mailbox/Email private data
