+++ imap/command.c
@@ -43,7 +43,10 @@
 #include "email/lib.h"
 #include "core/lib.h"
 #include "conn/lib.h"
+#include "adata.h"
+#include "edata.h"
 #include "init.h"
+#include "mdata.h"
 #include "message.h"
 #include "msn.h"
 #include "mutt_account.h"
