+++ imap/imap.c
@@ -46,11 +46,14 @@
 #include "lib.h"
 #include "bcache/lib.h"
 #include "pattern/lib.h"
+#include "adata.h"
 #include "auth.h"
 #include "command_parse.h"
 #include "commands.h"
+#include "edata.h"
 #include "hook.h"
 #include "init.h"
+#include "mdata.h"
 #include "message.h"
 #include "msn.h"
 #include "mutt_globals.h"
